// typescript/models/Coords.ts
/**
 * TravelTime API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.2.3
 * Contact: support@igeolise.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class Coords {
    'lat': number;
    'lng': number;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "lat",
            "baseName": "lat",
            "type": "number",
            "format": "double"
        },
        {
            "name": "lng",
            "baseName": "lng",
            "type": "number",
            "format": "double"
        }    ];

    static getAttributeTypeMap() {
        return Coords.attributeTypeMap;
    }
    
    public constructor() {
    }
}