// typescript/models/RequestLevelOfDetail.ts
/**
 * TravelTime API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.2.3
 * Contact: support@igeolise.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class RequestLevelOfDetail {
    'scaleType': RequestLevelOfDetailScaleTypeEnum;
    'level': RequestLevelOfDetailLevelEnum;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "scaleType",
            "baseName": "scale_type",
            "type": "RequestLevelOfDetailScaleTypeEnum",
            "format": ""
        },
        {
            "name": "level",
            "baseName": "level",
            "type": "RequestLevelOfDetailLevelEnum",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return RequestLevelOfDetail.attributeTypeMap;
    }
    
    public constructor() {
    }
}


export type RequestLevelOfDetailScaleTypeEnum = "simple" ;
export type RequestLevelOfDetailLevelEnum = "lowest" | "low" | "medium" | "high" | "highest" ;