// typescript/models/RequestSupportedLocations.ts
/**
 * TravelTime API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.2.3
 * Contact: support@igeolise.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestLocation } from './RequestLocation';
import { HttpFile } from '../http/http';

export class RequestSupportedLocations {
    'locations': Array<RequestLocation>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "locations",
            "baseName": "locations",
            "type": "Array<RequestLocation>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return RequestSupportedLocations.attributeTypeMap;
    }
    
    public constructor() {
    }
}